requirements.txt
future~=0.18.3
messagebox~=0.1.0
Pillow~=9.5.0